Make some defensive fixes to Snake logic.
SnakeBoss3 was glitching.  Circumstances seemed to be player dying at
the same time as shooting the snake, so the snake has to shrink then
change to an egg.  In one case the change to an egg wasn't happening; in
the other the egg wasn't at a grid junction, which resulted in the
reborn snake being unable to move.

The glitches are difficult to reproduce, so can't say for sure that
they're fixed now.